requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
cycler==0.10.0
dateparser==1.1.0
idna==3.3
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.4.3
multidict==5.2.0
numpy==1.21.3
pandas==1.3.4
patsy==0.5.2
Pillow==8.4.0
pyparsing==3.0.0
python-binance==1.0.15
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2021.10.23
requests==2.26.0
scikit-learn==1.0
scipy==1.7.1
six==1.16.0
sklearn==0.0
statsmodels==0.13.0
threadpoolctl==3.0.0
typing-extensions==3.10.0.2
tzdata==2021.4
tzlocal==4.0.1
ujson==4.2.0
urllib3==1.26.7
websockets==9.1
yarl==1.7.0
